Add a check on the subpixel search to exclude lousy fits.  This is key.
-Bring residual up from fit, and normalize it to span of intensity values
determined from the histogram analysis.  Threshold of 0.75 is pretty
arbitrary.   Based on analysis of 1 data set, 0.5 is too strict, but 1.0
is too easy.  Most of the obvious failures would be caught by 0.9, but a
few would sneak by.
-Badly fitted points were really dragging down the quality of the
calibration, and selfcal can't fix higher order distortion.
-Fixed a failure case in histogram analysis when we end up with imaginary
roots.  Is there a better way to handle this?